__init__ in billing_report, changes made to prepaid_billing
